# swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "This is Swagger UI environment generated for the TMF Product Catalog Management specification",
    "version" : "2.0",
    "title" : "Product Catalog Management"
  },
  "host" : "env-0693795.jelastic.servint.net",
  "basePath" : "/DSProductCatalog/api/catalogManagement/v2",
  "schemes" : [ "http" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
  "paths" : {
    "/category" : {
      "get" : {
        "tags" : [ "category" ],
        "summary" : "List or find 'Category' objects",
        "operationId" : "listCategory",
        "parameters" : [ {
          "name" : "fields",
          "in" : "query",
          "description" : "Comma separated properties to display in response",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Category"
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "category" ],
        "summary" : "Creates a 'Category'",
        "operationId" : "createCategory",
        "parameters" : [ {
          "in" : "body",
          "name" : "category",
          "description" : "The Category to be created",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category_Create"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Created",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/category/{id}" : {
      "get" : {
        "tags" : [ "category" ],
        "summary" : "Retrieves a 'Category' by Id",
        "operationId" : "retrieveCategory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Category",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "fields",
          "in" : "query",
          "description" : "Comma separated properties to display in response",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Ok",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Category"
              }
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "put" : {
        "tags" : [ "category" ],
        "summary" : "Updates a 'Category' by Id",
        "operationId" : "updateCategory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Category",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "category",
          "description" : "The Category to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category_Update"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      },
      "patch" : {
        "tags" : [ "category" ],
        "summary" : "Updates partially a 'Category' by Id",
        "operationId" : "updateCategory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Identifier of the Category",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "category",
          "description" : "The Category to be updated",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Category_Update"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Updated",
            "schema" : {
              "$ref" : "#/definitions/Category"
            }
          },
          "400" : {
            "description" : "Bad Request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "404" : {
            "description" : "Not Found",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
...
    "/hub" : {
      "get" : {
        "tags" : [ "hub" ],
        "summary" : "hubFind",
        "description" : "",
        "operationId" : "hubFind",
        "parameters" : [ ],
        "responses" : {
          "200" : {
            "description" : "List of Hub",
            "schema" : {
              "type" : "array",
              "items" : {
                "$ref" : "#/definitions/Hub"
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "hub" ],
        "summary" : "hubCreate",
        "description" : "",
        "operationId" : "hubCreate",
        "parameters" : [ {
          "in" : "body",
          "name" : "hub",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/Hub"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "Hub",
            "schema" : {
              "$ref" : "#/definitions/Hub"
            }
          }
        }
      }
    },
    "/hub/{hubId}" : {
      "get" : {
        "tags" : [ "hub" ],
        "summary" : "hubGet",
        "description" : "",
        "operationId" : "hubGet",
        "parameters" : [ {
          "name" : "hubId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Hub",
            "schema" : {
              "$ref" : "#/definitions/Hub"
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "hub" ],
        "summary" : "hubDelete",
        "description" : "",
        "operationId" : "hubDelete",
        "parameters" : [ {
          "name" : "hubId",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "204" : {
            "description" : "Success"
          }
        }
      }
    }
  },
  "definitions" : {
    "Category" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string",
          "description" : "Unique identifier of the category"
        },
        "href" : {
          "type" : "string",
          "description" : "Reference of the catalog"
        },
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date and time of the last update"
        },
        "version" : {
          "type" : "string",
          "description" : "Category version"
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the category is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "parentId" : {
          "type" : "string",
          "description" : "Unique identifier of the parent category"
        },
        "isRoot" : {
          "type" : "boolean",
          "description" : "If true, this Boolean indicates that the category is a root of categories"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the category"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the category"
        }
      },
      "description" : "The category resource is used to group product offerings, service and resource candidates in logical containers. Categories can contain other categories and/or product offerings, resource or service candidates."
    },
    "Category_Create" : {
      "type" : "object",
      "required" : [ "name", "parentId" ],
      "properties" : {
        "lastUpdate" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "Date and time of the last update"
        },
        "version" : {
          "type" : "string",
          "description" : "Category version"
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the category is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "parentId" : {
          "type" : "string",
          "description" : "Unique identifier of the parent category"
        },
        "isRoot" : {
          "type" : "boolean",
          "description" : "If true, this Boolean indicates that the category is a root of categories"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the category"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the category"
        }
      },
      "description" : "The category resource is used to group product offerings, service and resource candidates in logical containers. Categories can contain other categories and/or product offerings, resource or service candidates.\nSkipped properties: id,href"
    },
    "Category_Update" : {
      "type" : "object",
      "properties" : {
        "version" : {
          "type" : "string",
          "description" : "Category version"
        },
        "lifecycleStatus" : {
          "type" : "string",
          "description" : "Used to indicate the current lifecycle status"
        },
        "validFor" : {
          "description" : "The period for which the category is valid",
          "$ref" : "#/definitions/TimePeriod"
        },
        "parentId" : {
          "type" : "string",
          "description" : "Unique identifier of the parent category"
        },
        "isRoot" : {
          "type" : "boolean",
          "description" : "If true, this Boolean indicates that the category is a root of categories"
        },
        "name" : {
          "type" : "string",
          "description" : "Name of the category"
        },
        "description" : {
          "type" : "string",
          "description" : "Description of the category"
        }
      },
      "description" : "The category resource is used to group product offerings, service and resource candidates in logical containers. Categories can contain other categories and/or product offerings, resource or service candidates.\nSkipped properties: id,href,id,lastUpdate"
    },
...
    "TimePeriod" : {
      "type" : "object",
      "properties" : {
        "startDateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "An instant of time, starting at the TimePeriod"
        },
        "endDateTime" : {
          "type" : "string",
          "format" : "date-time",
          "description" : "An instant of time, ending at the TimePeriod."
        }
      },
      "description" : "A base / value business entity used to represent a period of time between two timepoints."
    },
...
    "Hub" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "callback" : {
          "type" : "string"
        },
        "query" : {
          "type" : "string"
        }
      }
    },
    "Error" : {
      "required" : [ "code", "message" ],
      "properties" : {
        "code" : {
          "type" : "integer",
          "description" : "An integer coding the error type. This is given to caller so he can translate them if required."
        },
        "message" : {
          "type" : "string",
          "description" : "A short localized string that describes the error."
        },
        "description" : {
          "type" : "string",
          "description" : "(optional) A long localized error description if needed. It can contain precise information about which parameter is missing, or what are the identifier acceptable values."
        },
        "infoURL" : {
          "type" : "string",
          "description" : "(optional) A URL to online documentation that provides more information about the error."
        }
      }
    }
  }
}